It's not actually pushing to server I don't think
I can see b/c console.log stuff that it’s reading what’s in the input
and it says it’s pushing to array, and it does for half a second but
not permanently. I’m not sure how to get it to write perm. I’m using
$update which is what edit-story uses and it works, rewrites the db
stuff, so…